backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.5.0,<3
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
gitpython>=3.1.40
markdown>=3.5.2
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.1.0
httpx>=0.27.0
requests>=2.32.0
google-generativeai>=0.4.1
langchain>=0.2.0
langchain-google-genai>=0.0.9
transformers>=4.40.0
torch>=2.1.0
python-jose[cryptography]>=3.3.0

